Moved term constants and helper functions to shared/terms
