+++ client/plots/summary.js
@@ -7,7 +7,7 @@ import { getDefaultViolinSettings } from './violin.js'
 import { getDefaultBarSettings } from './barchart.js'
 import { getDefaultScatterSettings } from './sampleScatter.js'
 import { Tabs } from '../dom/toggleButtons'
-import { isNumericTerm } from '#shared/common.js'
+import { isNumericTerm } from '../shared/terms'
 
 //import {  } from ''
