+++ client/plots/violin.js
@@ -4,7 +4,8 @@ import setViolinRenderer from './violin.renderer'
 import htmlLegend from '../dom/html.legend'
 import { fillTermWrapper } from '#termsetting'
 import { setInteractivity } from './violin.interactivity'
-import { plotColor, isNumericTerm } from '#shared/common'
+import { plotColor } from '#shared/common'
+import { isNumericTerm } from '#shared/terms'
 
 /*
 when opts.mode = 'minimal', a minimal violin plot will be rendered that will have a single term and minimal features (i.e. no controls, legend, labels, brushing, transitions, etc.)
