+++ client/termsetting/handlers/geneExpression.ts
@@ -1,6 +1,6 @@
 import { NumericQ } from '../../shared/types/terms/numeric'
 import { VocabApi } from '../../shared/types/index'
-import { GeneExpressionTW } from 'shared/types/terms/geneExpression.js'
+import { GeneExpressionTW } from '../../shared/types/terms/geneExpression.js'
 
 /*
 Routes numeric terms to their respective subhandlers. Functions follow the same naming convention as the other handler files and returns the results.
