+++ server/shared/common.js
@@ -1094,55 +1094,3 @@ export const mutationClasses = Object.values(mclass)
 export const CNVClasses = Object.values(mclass)
 	.filter(m => m.dt == dtcnv)
 	.map(m => m.key)
-
-//Term types should be used gradually using these constants instead of hardcoding the values,
-// eg: type == TermTypes.CATEGORICAL instead of type == 'categorical'
-export const TermTypes = {
-	GENE_VARIANT: 'geneVariant',
-	GENE_EXPRESSION: 'geneExpression',
-	CATEGORICAL: 'categorical',
-	INTEGER: 'integer',
-	FLOAT: 'float',
-	SNP_LIST: 'snplst',
-	SNP_LOCUS: 'snplocus',
-	CONDITION: 'condition',
-	SURVIVAL: 'survival',
-	SAMPLELST: 'samplelst'
-}
-
-export const TermTypeGroups = {
-	DICTIONARY_VARIABLES: 'Dictionary Variables',
-	MUTATION_CNV_FUSION: 'Mutation/CNV/Fusion',
-	VARIANT_GENOTYPE: 'Variant Genotype',
-	DNA_METHYLATION: 'DNA Methylation',
-	GENE_DEPENDENCY: 'Gene Dependency',
-	GENE_EXPRESSION: 'Gene Expression',
-	PROTEIN_EXPRESSION: 'Protein Expression',
-	SPLICE_JUNCTION: 'Splice Junction',
-	METABOLITE: 'Metabolite',
-	GSEA: 'GSEA',
-	MUTATION_SIGNATURE: 'Mutation Signature',
-	SNP_LIST: 'SNP List',
-	SNP_LOCUS: 'SNP Locus'
-}
-
-export function isNumericTerm(term) {
-	if (!term) return false
-	return term.type == TermTypes.INTEGER || term.type == TermTypes.FLOAT || term.type == TermTypes.GENE_EXPRESSION
-}
-
-const nonDictTypes = new Set([
-	TermTypes.SNP_LIST,
-	TermTypes.SNP_LOCUS,
-	TermTypes.GENE_EXPRESSION,
-	TermTypes.GENE_VARIANT
-])
-
-export function isDictionaryType(type) {
-	return !isNonDictionaryType(type)
-}
-
-export function isNonDictionaryType(type) {
-	if (!type) throw new Error('Type is not defined')
-	return nonDictTypes.has(type)
-}
