+++ server/src/termdb.violin.js
@@ -5,7 +5,7 @@ import { createCanvas } from 'canvas'
 import { getBinsDensity } from '../../server/shared/violin.bins'
 import summaryStats from '../shared/descriptive.stats'
 import { getOrderedLabels } from './termdb.barchart'
-import { isNumericTerm } from '#shared/common.js'
+import { isNumericTerm } from '#shared/terms.js'
 /*
 q = {
   getViolinPlotData: '1',
